+++ kernel/kernel-api/src/test/scala/io/delta/kernel/internal/tablefeatures/TableFeaturesSuite.scala
@@ -15,24 +15,20 @@
  */
 package io.delta.kernel.internal.tablefeatures
 
-import io.delta.kernel.data.{ArrayValue, ColumnVector, MapValue}
 import io.delta.kernel.exceptions.KernelException
 import io.delta.kernel.internal.actions.{Format, Metadata, Protocol}
 import io.delta.kernel.internal.tablefeatures.TableFeatures.{TABLE_FEATURES, validateWriteSupportedTable}
-import io.delta.kernel.internal.util.InternalUtils.singletonStringColumnVector
-import io.delta.kernel.internal.util.VectorUtils.stringVector
-import io.delta.kernel.types._
+import io.delta.kernel.test.TableMetadataTestUtils
 import org.scalatest.funsuite.AnyFunSuite
 
-import java.util.stream.Collectors
 import java.util.stream.Collectors.toList
-import java.util.{Collections, Optional}
+import java.util.{Collections}
 import scala.collection.JavaConverters._
 
 /**
  * Suite that tests Kernel throws error when it receives a unsupported protocol and metadata
  */
-class TableFeaturesSuite extends AnyFunSuite {
+class TableFeaturesSuite extends AnyFunSuite with TableMetadataTestUtils {
 
   /////////////////////////////////////////////////////////////////////////////////////////////////
   // Tests for [[TableFeature]] implementations                                                  //
@@ -258,95 +254,5 @@ class TableFeaturesSuite extends AnyFunSuite {
     intercept[KernelException] {
       validateWriteSupportedTable(protocol, metadata, "/test/table")
     }
-  }
-
-  def createTestProtocol(minWriterVersion: Int, writerFeatures: String*): Protocol = {
-    new Protocol(
-      // minReaderVersion - it doesn't matter as the read fails anyway before the writer check
-      0,
-      minWriterVersion,
-      // reader features - it doesn't matter as the read fails anyway before the writer check
-      Collections.emptyList(),
-      writerFeatures.toSeq.asJava
-    )
-  }
-
-  def testMetadata(
-      includeInvaraint: Boolean = false,
-      includeTimestampNtzTypeCol: Boolean = false,
-      includeVariantTypeCol: Boolean = false,
-      includeGeneratedColumn: Boolean = false,
-      includeIdentityColumn: Boolean = false,
-      tblProps: Map[String, String] = Map.empty): Metadata = {
-    val testSchema = createTestSchema(
-      includeInvaraint,
-      includeTimestampNtzTypeCol,
-      includeVariantTypeCol,
-      includeGeneratedColumn,
-      includeIdentityColumn)
-    new Metadata(
-      "id",
-      Optional.of("name"),
-      Optional.of("description"),
-      new Format("parquet", Collections.emptyMap()),
-      testSchema.toJson,
-      testSchema,
-      new ArrayValue() { // partitionColumns
-        override def getSize = 1
-
-        override def getElements: ColumnVector = singletonStringColumnVector("c3")
-      },
-      Optional.empty(),
-      new MapValue() { // conf
-        override def getSize = tblProps.size
-        override def getKeys: ColumnVector = stringVector(tblProps.toSeq.map(_._1).asJava)
-        override def getValues: ColumnVector = stringVector(tblProps.toSeq.map(_._2).asJava)
-      }
-    )
-  }
-
-  def createTestSchema(
-      includeInvariant: Boolean = false,
-      includeTimestampNtzTypeCol: Boolean = false,
-      includeVariantTypeCol: Boolean = false,
-      includeGeneratedColumn: Boolean = false,
-      includeIdentityColumn: Boolean = false): StructType = {
-    var structType = new StructType()
-      .add("c1", IntegerType.INTEGER)
-      .add("c2", StringType.STRING)
-    if (includeInvariant) {
-      structType = structType.add(
-        "c3",
-        TimestampType.TIMESTAMP,
-        FieldMetadata.builder()
-          .putString("delta.invariants", "{\"expression\": { \"expression\": \"x > 3\"} }")
-          .build())
-    }
-    if (includeTimestampNtzTypeCol) {
-      structType = structType.add("c4", TimestampNTZType.TIMESTAMP_NTZ)
-    }
-    if (includeVariantTypeCol) {
-      structType = structType.add("c5", VariantType.VARIANT)
-    }
-    if (includeGeneratedColumn) {
-      structType = structType.add(
-        "c6",
-        IntegerType.INTEGER,
-        FieldMetadata.builder()
-          .putString("delta.generationExpression", "{\"expression\": \"c1 + 1\"}")
-          .build())
-    }
-    if (includeIdentityColumn) {
-      structType = structType.add(
-        "c7",
-        IntegerType.INTEGER,
-        FieldMetadata.builder()
-          .putLong("delta.identity.start", 1L)
-          .putLong("delta.identity.step", 2L)
-          .putBoolean("delta.identity.allowExplicitInsert", true)
-          .build())
-    }
-
-    structType
   }
 }
